+++ Services/Orders/Cart/ICartService.cs
@@ -8,7 +8,7 @@ public interface ICartService
     Task<CartMutationResult> AddToCartAsync(
         string sessionId,
         int productId,
-        int quantity = 1,
+        decimal quantity = 1m,
         int? variantId = null);
     Task<CartMutationResult> AddItemsToCartAsync(
         string sessionId,
@@ -16,11 +16,11 @@ public interface ICartService
     Task<CartMutationResult> AddComboToCartAsync(string sessionId, int comboId);
     Task<CartMutationResult> UpdateComboQuantityAsync(string sessionId, int cartGroupId, int quantity);
     Task RemoveComboAsync(string sessionId, int cartGroupId);
-    Task UpdateQuantityAsync(string sessionId, int cartItemId, int quantity);
+    Task UpdateQuantityAsync(string sessionId, int cartItemId, decimal quantity);
     Task RemoveFromCartAsync(string sessionId, int cartItemId);
     Task<CartViewModel> RepriceForCheckoutAsync(string sessionId);
     Task ClearCartAsync(string sessionId);
-    Task<int> GetCartCountAsync(string sessionId);
+    Task<decimal> GetCartCountAsync(string sessionId);
     Task<CouponApplyResult> ApplyCouponAsync(string sessionId, string couponCode);
     Task RemoveCouponAsync(string sessionId);
 }
