+++ Services/Pricing/Coupons/ICouponService.cs
@@ -12,8 +12,8 @@ public interface ICouponService
     Task<(bool Success, string? Error)> DeleteAsync(int id);
 
     // Kiểm tra mã và tính discount. Trả về kết quả áp dụng
-    Task<CouponApplyResult> ApplyCouponAsync(string code, decimal subtotal, int itemCount);
+    Task<CouponApplyResult> ApplyCouponAsync(string code, decimal subtotal, decimal itemCount);
 
     // Lấy danh sách coupon đang hoạt động với trạng thái đủ điều kiện
-    Task<List<CouponEligibilityResult>> GetAvailableCouponsAsync(decimal subtotal, int itemCount);
+    Task<List<CouponEligibilityResult>> GetAvailableCouponsAsync(decimal subtotal, decimal itemCount);
 }
